refactor: consolidate chunk stat creation (#8822)
- remove `create_basic_summary`, everyone shall use the more powerful
  `create_chunk_statistics`
- make column summaries in `create_chunk_statistics` optional to avoid
  that users allocate standard-size hashmaps just for nothing
- make row count `usize` (instead of `u64`) to avoid a few casts
- make row count optional to better deal with streaming data

Found while working on #8350.

+++ ingester/src/buffer_tree/partition.rs
@@ -337,7 +337,7 @@ impl PartitionData {
         // the row count reported by the rows() method.
         //
         // The row count is never affected by projection.
-        debug_assert_eq!(q.num_rows(), self.rows() as u64);
+        debug_assert_eq!(q.num_rows(), self.rows());
 
         // Invariant: the timestamp min/max MUST match the values reported by
         // timestamp_stats(), iff the projection contains the "time" column.
